+++ src/util/protocol.ts
@@ -98,6 +98,7 @@ export {
   ImplementationRequest,
   DocumentFormattingRequest,
   DocumentRangeFormattingRequest,
+  DocumentRangesFormattingRequest,
   DocumentOnTypeFormattingRequest,
   FoldingRangeRequest,
   DidChangeWatchedFilesNotification,